+++ apps/velocity/pages/analytics.tsx
@@ -18,6 +18,7 @@ import {
 import {
   Card,
   MetricCard,
+  Text,
   mq,
 } from '@terenceodonoghue/react-components/velocity';
 import {
@@ -44,24 +45,29 @@ const AnalyticsPage: NextPage = () => {
       }}
     >
       {payload.map((entry) => (
-        <li
+        <Text
           css={mq({
-            alignItems: 'center',
-            color: color.neutral[600],
             display: 'flex',
+            alignItems: 'center',
             margin: ['0 6px', '0 12px'],
             '&::before': {
-              content: '"\u25CF"',
-              color: entry.color,
+              content: '""',
+              borderRadius: '50%',
+              backgroundColor: entry.color,
               display: 'inline-block',
               fontSize: '1.5em',
               marginRight: 8,
+              height: 10,
+              width: 10,
             },
           })}
+          align="center"
+          as="li"
           key={entry.dataKey}
+          variant="p2"
         >
           {entry.value}
-        </li>
+        </Text>
       ))}
     </ul>
   );
@@ -211,6 +217,7 @@ const AnalyticsPage: NextPage = () => {
                 />
                 <XAxis
                   dataKey="name"
+                  fontSize={15}
                   interval="preserveStartEnd"
                   stroke={color.secondary}
                   tick={{ fill: color.neutral[500] }}
@@ -218,6 +225,7 @@ const AnalyticsPage: NextPage = () => {
                   tickMargin={8.5}
                 />
                 <YAxis
+                  fontSize={15}
                   stroke={color.secondary}
                   tickFormatter={yTickFormatter}
                   tickLine={false}
@@ -365,6 +373,7 @@ const AnalyticsPage: NextPage = () => {
                 />
                 <XAxis
                   dataKey="name"
+                  fontSize={15}
                   interval="preserveStartEnd"
                   padding={{ left: 17, right: 16 }}
                   stroke={color.secondary}
@@ -373,6 +382,7 @@ const AnalyticsPage: NextPage = () => {
                   tickMargin={7.5}
                 />
                 <YAxis
+                  fontSize={15}
                   tickLine={false}
                   tickMargin={1}
                   stroke={color.secondary}