+++ apps/velocity/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @terenceodonoghue/velocity
 
+## 1.0.26
+
+### Patch Changes
+
+- Updated dependencies [f5609d8]
+- Updated dependencies [1167a2b]
+  - @terenceodonoghue/react-components@1.3.1
+
+## 1.0.25
+
+### Patch Changes
+
+- Updated dependencies [c84e3dd]
+- Updated dependencies [bfc44aa]
+  - @terenceodonoghue/react-components@1.3.0
+
 ## 1.0.24
 
 ### Patch Changes