+++ packages/react-components/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @terenceodonoghue/react-components
 
+## 1.3.1
+
+### Patch Changes
+
+- f5609d8: Fix font on Box, Card components
+- 1167a2b: Fix font on Badge, Pill components
+
+## 1.3.0
+
+### Minor Changes
+
+- c84e3dd: Remove core components
+
+### Patch Changes
+
+- bfc44aa: Add global `styles` prop to ThemeProvider
+
 ## 1.2.0
 
 ### Minor Changes